upload_document_to_azure_blob/requirements.txt
azure-common==1.1.27
azure-core==1.17.0
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==12.8.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Flask==2.0.1
Flask-Cors==3.0.10
futures==3.1.1
idna==3.2
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msrest==0.6.21
oauthlib==3.1.1
pycparser==2.20
python-dateutil==2.8.2
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.6
waitress==2.0.0
Werkzeug==2.0.1
